refactor(events): move business logic out of events/service/__init__.py (#412)
Move `validate_feedback_requires_evaluation` and `update_organization_questionnaire`
to `events/service/event_questionnaire_service.py`, and `update_db_instance` (a
generic ORM helper) to `common/utils.py`. The `events/service/__init__.py` is
now a pure re-export module with an explicit `__all__`, so existing
`from events.service import ...` call sites continue to work.
